requirements.txt
Flask==2.2.5
Flask-Cors==5.0.0
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.1
Pillow==10.2.0
segmentation-models-pytorch==0.3.4
tensorflow==2.16.2
torch==2.2.2
